Fix review findings: ZIP offset overflow, OOB triangle indices, unclosed SVG paths, MSVC shadow warning
- ZipReader: widen localOffset+30 to size_t before comparing against
  data.size() — in 32-bit space a malicious/malformed central directory
  entry with localOffset near 0xFFFFFFFF wraps back into range, defeating
  the bounds check and reading the ZIP local-file-header signature out of
  bounds.
- ThreeMfLoader/AmfLoader: bounds-check triangle v1/v2/v3 against the
  enclosing object's own vertex count, not just non-negativity — an
  out-of-range index (e.g. a corrupt file, or one with far fewer <vertex>
  entries than a <triangle> references) is now skipped instead of feeding
  an out-of-bounds index into the position buffer downstream.
- SvgLoader: a path's subpath only counts as closed if it actually hit a
  Z/z command — previously any subpath with >=3 points was treated as
  closed regardless, silently importing e.g. "M0,0 L4,0 L4,4 L0,4" (no
  trailing Z) as if it were closed.
- Fix the actual Windows CI failure: MSVC's /WX treats C4457 (local
  variable hides a parameter) as an error — DxfLoader.cpp and
  SvgLoader.cpp both had a `for (auto& path : closedPaths)` shadowing the
  function's own `path` parameter; renamed to `contour`.

New regression tests for each of the four fixes.

+++ src/import/ZipReader.cpp
@@ -119,13 +119,19 @@ ZipExtractResult zipExtractBySuffix(const std::filesystem::path& archivePath,
         return out;
     }
 
-    if (localOffset + 30 > data.size() || readU32(data, localOffset) != kLocalSig) {
+    // Widen to size_t before adding: localOffset is uint32_t, and a
+    // malformed/adversarial central directory entry could set it near
+    // 0xFFFFFFFF, which would wrap back into range in 32-bit arithmetic and
+    // defeat this bounds check entirely.
+    if (static_cast<std::size_t>(localOffset) + 30 > data.size() ||
+        readU32(data, localOffset) != kLocalSig) {
         out.error = "malformed ZIP local file header";
         return out;
     }
     uint16_t localNameLen = readU16(data, localOffset + 26);
     uint16_t localExtraLen = readU16(data, localOffset + 28);
-    std::size_t dataStart = localOffset + 30 + localNameLen + localExtraLen;
+    std::size_t dataStart =
+        static_cast<std::size_t>(localOffset) + 30 + localNameLen + localExtraLen;
     if (dataStart + compSize > data.size()) {
         out.error = "malformed ZIP entry (compressed data extends past end of file)";
         return out;
